Usando la matriz del otro grupo para clasificar

+++ WorldDomination/src/Parseador.h
@@ -14,7 +14,7 @@
 
 struct review {
     std::string id;
-    int sentimiento;
+    double sentimiento;
     std::string texto;
   };
